Moves RestartData source and header files to /src and /include
