+++ src/disturb/WildFire.h
@@ -10,7 +10,7 @@
 #include "../../include/EnvData.h"
 #include "../../include/FireData.h"
 #include "../../include/BgcData.h"
-#include "../data/RestartData.h"
+#include "../../include/RestartData.h"
 
 #include "../inc/errorcode.h"
 #include "../inc/timeconst.h"
