+++ src/ecodomain/Ground.h
@@ -27,7 +27,7 @@ using namespace std;
 #include "../../include/BgcData.h"
 #include "../../include/CohortData.h"
 #include "../../include/FireData.h"
-#include "../data/RestartData.h"
+#include "../../include/RestartData.h"
 
 #include "../inc/parameters.h"
